requirements.txt
Django==1.8.5
djangorestframework==3.2.3
wheel==0.24.0
